create a votemonitor.domain project and move the datacontext there
