+++ src/Vote.Monitor.Api/Vote.Monitor.Api.csproj
@@ -24,9 +24,9 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Vote.Monitor.Core\Vote.Monitor.Core.csproj" />
-    <ProjectReference Include="..\Vote.Monitor.Feature.Example\Vote.Monitor.Feature.Example.csproj" />
+    <ProjectReference Include="..\Vote.Monitor.Domain\Vote.Monitor.Domain.csproj" />
     <ProjectReference Include="..\Vote.Monitor.Feature.PollingStation\Vote.Monitor.Feature.PollingStation.csproj" />
-    <ProjectReference Include="..\Vote.Monitor.Infra\Vote.Monitor.Infra.csproj" />
+    <ProjectReference Include="..\Vote.Monitor.Infra\Vote.Monitor.Domain.csproj" />
   </ItemGroup>
 
 </Project>
