+++ src/Vote.Monitor.Domain/Models/TagModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace Vote.Monitor.Feature.PollingStation.Models;
+namespace Vote.Monitor.Domain.Models;
 public  class TagModel
 {
     [Key] public int Id { get; set; }
