+++ src/Vote.Monitor.Domain/Models/TagModelExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace Vote.Monitor.Feature.PollingStation.Models;
+﻿namespace Vote.Monitor.Domain.Models;
 
 public static class TagModelExtensions
 {
