+++ src/Vote.Monitor.Feature.PollingStation/GetAllPollingStations/GetAllPollingStationsEndpoint.cs
@@ -1,6 +1,5 @@
-﻿using System.Text.Json;
-using FastEndpoints;
-using Vote.Monitor.Feature.PollingStation.Models;
+﻿using FastEndpoints;
+using Vote.Monitor.Domain.Models;
 using Vote.Monitor.Feature.PollingStation.Repositories;
 
 namespace Vote.Monitor.Feature.PollingStation.GetAllPollingStations;
