+++ src/Vote.Monitor.Feature.PollingStation/Repositories/IPollingStationRepository.cs
@@ -1,5 +1,5 @@
-﻿using Vote.Monitor.Feature.PollingStation.Models;
-using Vote.Monitor.Core;
+﻿using Vote.Monitor.Core;
+using Vote.Monitor.Domain.Models;
 
 namespace Vote.Monitor.Feature.PollingStation.Repositories;
 internal interface IPollingStationRepository:IRepository<PollingStationModel>
