--- tests/Vote.Monitor.Feature.Example.IntegrationTests/GreetTests/Fakes.cs
@@ -1,12 +1,0 @@
-﻿using Bogus;
-using Greet = Vote.Monitor.Feature.Example.Greet;
-
-namespace Tests.GreetTests;
-
-static class Fakes
-{
-    public static Greet.Request GreetRequest(this Faker f) => new()
-    {
-        Name = f.Name.FullName()
-    };
-}
