+++ tests/Vote.Monitor.Feature.Example.IntegrationTests/Usings.cs
@@ -1,6 +1,1 @@
-﻿global using FastEndpoints;
-global using FastEndpoints.Testing;
-global using FluentAssertions;
-global using Xunit;
-global using Xunit.Abstractions;
-global using Xunit.Priority;
+﻿
