+++ tests/Vote.Monitor.Feature.Example.IntegrationTests/Vote.Monitor.Feature.PollingStation.IntegrationTests.csproj
