+++ tests/Vote.Monitor.Feature.Example.UnitTests/Usings.cs
@@ -1,3 +1,1 @@
-﻿global using FluentAssertions;
-global using Xunit;
-global using FluentValidation.TestHelper;
+﻿
